requirements.txt
# Core framework (obbligatorio)
datapizza-ai
openai

# Vector store
qdrant-client

# Document parsing
docling
beautifulsoup4
pandas

# Hybrid search
rank-bm25

# Embedding locale su GPU (gratis, RTX A1000 / Kaggle T4)
sentence-transformers
torch
einops

# Demo
streamlit==1.54.0
chainlit

# Utils
python-dotenv
tqdm
httpx
